Mis read the guide the first time https://community.jboss.org/wiki/WildFlyNamingPolicy/ so the naming is changes to match org.wildfly.extras according to the guide
