+++ arqullian-integration-tests/pom.xml
@@ -20,8 +20,8 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.wildfly.extension.db_bootstrap</groupId>
-        <artifactId>wildfly-db-bootstrap-subsystem-parent</artifactId>
+        <groupId>org.wildfly.extras.db_bootstrap</groupId>
+        <artifactId>db-bootstrap-subsystem-parent</artifactId>
         <version>1.0.5</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
@@ -53,8 +53,8 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.wildfly.extension.db_bootstrap</groupId>
-            <artifactId>wildfly-db-bootstrap</artifactId>
+            <groupId>org.wildfly.extras.db_bootstrap</groupId>
+            <artifactId>db-bootstrap</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
