+++ extension/src/main/java/org/wildfly/extras/db_bootstrap/DbBootstrapDetectorAdd.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.extension.db_bootstrap;
+package org.wildfly.extras.db_bootstrap;
 
 import java.util.List;
 
@@ -26,25 +26,35 @@ import org.jboss.as.server.DeploymentProcessorTarget;
 import org.jboss.as.server.deployment.Phase;
 import org.jboss.dmr.ModelNode;
 import org.jboss.msc.service.ServiceController;
-
-public class DbBootstrapSubsystemAdd extends AbstractBoottimeAddStepHandler {
+/**
+ * @author Frank Vissing
+ * @author Flemming Harms
+ */
+final class DbBootstrapDetectorAdd extends AbstractBoottimeAddStepHandler {
 
     @Override
     protected void populateModel(ModelNode operation, ModelNode model) throws OperationFailedException {
-        //noting to do as we do not have any attributes in model
-        model.get(DbBootstrapExtension.EXTENSION_TYPE).setEmptyObject();
+        DbBootstrapDetectorResourceDefinition.FILENAME.validateAndSet(operation, model);
+        DbBootstrapDetectorResourceDefinition.FILTER_ON_NAME.validateAndSet(operation, model);
     }
 
     @Override
     protected void performBoottime(OperationContext context, ModelNode operation, ModelNode model, ServiceVerificationHandler verificationHandler, List<ServiceController<?>> newControllers) throws OperationFailedException {
-        DbBootstrapLogger.ROOT_LOGGER.subsystemStarted();
+        final String filename = model.get(DbBootstrapExtension.FILENAME_ATTR).asString();
+        final String filterOnName =  model.get(DbBootstrapExtension.FILTER_ON_NAME_ATTR).asString();
 
         context.addStep(new AbstractDeploymentChainStep() {
-                   @Override
-                   protected void execute(DeploymentProcessorTarget processorTarget) {
-                       // Initialize the deployer chain
-                       processorTarget.addDeploymentProcessor(DbBootstrapExtension.SUBSYSTEM_NAME, Phase.PARSE, Phase.PARSE_WEB_DEPLOYMENT, new DbBootstrapSubsystemDetectorProcessor());
-                   }
-               }, OperationContext.Stage.RUNTIME);
+            @Override
+            protected void execute(DeploymentProcessorTarget processorTarget) {
+                DbBootstrapLogger.ROOT_LOGGER.tracef("%s:'%s' %s:'%s'",DbBootstrapExtension.FILENAME_ATTR,filename,DbBootstrapExtension.FILTER_ON_NAME_ATTR,filterOnName);
+                try {
+                    processorTarget.addDeploymentProcessor(DbBootstrapExtension.SUBSYSTEM_NAME, Phase.PARSE, Phase.PARSE_WEB_DEPLOYMENT,new DbBootstrapDetectorProcessor(filename,filterOnName));
+                } catch (Exception e) {
+                    DbBootstrapLogger.ROOT_LOGGER.error("Error in instanciating DbBootstraper add handler", e);
+                }
+
+
+            }
+        }, OperationContext.Stage.RUNTIME);
     }
 }
