+++ extension/src/main/java/org/wildfly/extras/db_bootstrap/DbBootstrapSubsystemParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.extension.db_bootstrap;
+package org.wildfly.extras.db_bootstrap;
 
 import static org.jboss.as.controller.parsing.ParseUtils.unexpectedElement;
