+++ extension/src/main/java/org/wildfly/extras/db_bootstrap/annotations/BootstrapDatabase.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.extension.db_bootstrap.annotations;
+package org.wildfly.extras.db_bootstrap.annotations;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
