+++ extension/src/main/java/org/wildfly/extras/db_bootstrap/annotations/UpdateSchema.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.extension.db_bootstrap.annotations;
+package org.wildfly.extras.db_bootstrap.annotations;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -22,10 +22,10 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Specifies this is a bootstrap method. This annotation applies to a method
+ * Specifies this is a update method. This annotation applies to a method
  * and is executed a part of the {@link org.jboss.as.server.deployment.Phase}.PARSE
  * and {@link org.jboss.as.server.deployment.Phase}.PARSE_EJB_DEPLOYMENT and is the
- * first phase of bootstrapping. For second phase see {@link UpdateSchema}
+ * second phase of bootstrapping and run after {@link BootstrapSchema}
  * <br><br>
  * If the parameter signature of the method is {@link org.hibernate.Session} the
  * hibernate session will be passed into the method call.
@@ -37,5 +37,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 @Documented
-public @interface BootstrapSchema {
+public @interface UpdateSchema {
 }
