+++ extension/src/test/resources/org/wildfly/extras/db_bootstrap/db_bootstrap-1.0.xml
