+++ integrate/pom.xml
@@ -20,13 +20,13 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.wildfly.extension.db_bootstrap</groupId>
-        <artifactId>wildfly-db-bootstrap-subsystem-parent</artifactId>
+        <groupId>org.wildfly.extras.db_bootstrap</groupId>
+        <artifactId>db-bootstrap-subsystem-parent</artifactId>
         <version>1.0.5</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>wildfly-extension-integrator</artifactId>
+    <artifactId>extension-integrator</artifactId>
 
     <name>Database bootstrap Extension</name>
     <packaging>jar</packaging>
@@ -36,7 +36,7 @@
         <version.rhino.js>1.7R2</version.rhino.js>
         <version.org.jboss.jandex>1.0.3.Final</version.org.jboss.jandex>
         <version.org.apache.ant>1.8.2</version.org.apache.ant>
-        <module.dest.directory>${wildfly.home}/modules/org/wildfly/extension/db_bootstrap/main</module.dest.directory>
+        <module.dest.directory>${wildfly.home}/modules/org/wildfly/extras/db_bootstrap/main</module.dest.directory>
         <!-- generated configs root dirs -->
         <generated.configs>${basedir}/target/generated-configs</generated.configs>
         <generated.configs.standalone>${generated.configs}/standalone/configuration</generated.configs.standalone>
@@ -123,8 +123,8 @@
     </build>
     <dependencies>
         <dependency>
-            <groupId>org.wildfly.extension.db_bootstrap</groupId>
-            <artifactId>wildfly-db-bootstrap</artifactId>
+            <groupId>org.wildfly.extras.db_bootstrap</groupId>
+            <artifactId>db-bootstrap</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
