+++ integrate/src/main/resources/standalone/configuration/standalone-db-bootstrap.xml
@@ -25,7 +25,7 @@
         <extension module="org.wildfly.extension.batch"/>
         <extension module="org.wildfly.extension.io"/>
         <extension module="org.wildfly.extension.undertow"/>
-        <extension module="org.wildfly.extension.db_bootstrap"/>
+        <extension module="org.wildfly.extras.db_bootstrap"/>
     </extensions>
     <management>
         <security-realms>
